+++ src/main/java/com/chelobyte/calculator/operation/Result.java
@@ -2,7 +2,7 @@ package com.chelobyte.calculator.operation;
 
 public class Result {
 
-    public static void printResult(int result) {
+    public static void printResult(double result) {
         System.out.println("The result is: " + result);
     }
